Update and rename install to install.sh

+++ install.sh
@@ -1,4 +1,17 @@
-#!/usr/bin/env bash
+hash git >/dev/null || {
+  echo "git not installed"
+  exit
+}
+
+cd $HOME
+
+if [ -d $HOME/.oh-my-zsh ]; then
+	git clone git://github.com/tinogomes/oh-my-zsh.git .oh-my-zsh
+fi
+
+if [ -d $HOME/rcfiles ]; then
+	git clone git@github.com:tinogomes/rcfiles.git
+fi
 
 ln -s $HOME/.rcfiles/gemrc $HOME/.gemrc
 ln -s $HOME/.rcfiles/gitattributes-global $HOME/.gitattributes
@@ -14,7 +27,9 @@ ln -s $HOME/.rcfiles/slate $HOME/.slate
 echo "Write on $HOME/.zshrc_local file your local settings"
 echo "Write here your local settings" >> $HOME/.zshrc_local
 
-mv $HOME/.zshrc $HOME/.zshrc_original
+if [ -f $HOME/.zshrc ]; then
+	mv $HOME/.zshrc $HOME/.zshrc_original
+fi
 
 case $1 in
 	ubuntu|linux) ln -s $HOME/.rcfiles/zshrc_ubuntu $HOME/.zshrc;;
